abstract out table and remove unnecessary files
