+++ app/news/page.js
@@ -1,37 +1,29 @@
-'use client'
+import { promises as fs } from 'fs';
+import Table from '../ui/table';
 
-import { useState, useEffect } from "react";
+export default async function News() {
+    const file = await fs.readFile(process.cwd() + '/app/news/data/test.json', 'utf8');
+    const data = JSON.parse(file);
 
-export default function News() {
-    const [data, setData] = useState({news: []});
+    const columns = [
+        {
+            key: "Title",
+            label: "Title",
+        },{
+            key: "Source",
+            label: "Source",
+        },{
+            key: "PublishedOn",
+            label: "Published On",
+        },{
+            key: "Description",
+            label: "Description",
+        }
+    ]
 
-    useEffect(() => {
-        console.log("fetching news now!")
-        fetch('/api/news')
-        .then(res => res.json())
-        .then(resData => {
-            console.log(resData.data);
-            setData(resData.data);
-        })
-    }, [])
-
+    const items = data.news;
+    
     return (
-        <div>
-            {
-                data.news.map(article => {
-                    return (
-                        <div key={article.Title}>
-                            <p>{article.Title}</p>
-                            <p>{article.Source}</p>
-                            <p>{article.Url}</p>
-                            <p>{article.PublishedOn}</p>
-                            <p>{article.Description}</p>
-                            <p>{article.Summary}</p>
-                            <p>{article.Categories.label}</p>
-                        </div>
-                    )
-                })
-            }
-        </div>
+        <Table columns={columns} items={items} />
     )
 }
